+++ src/FreezeableInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace League\Route;
+
+interface FreezeableInterface
+{
+    public function freeze(): void;
+    public function isFrozen(): bool;
+}