+++ src/Http/Exception/MethodNotAllowedException.php
@@ -9,17 +9,19 @@ use League\Route\Http;
 
 class MethodNotAllowedException extends Http\Exception
 {
-    /** @param array<string> $allowed */
+    /** @param array<string> $allowedMethods */
     public function __construct(
-        array $allowed = [],
+        private readonly array $allowedMethods = [],
         string $message = 'Method Not Allowed',
         ?Exception $previous = null,
         int $code = 0,
     ) {
-        $headers = [
-            'Allow' => implode(', ', $allowed),
-        ];
+        parent::__construct(405, $message, $previous, ['Allow' => implode(', ', $allowedMethods)], $code);
+    }
 
-        parent::__construct(405, $message, $previous, $headers, $code);
+    /** @return array<string> */
+    public function getAllowedMethods(): array
+    {
+        return $this->allowedMethods;
     }
 }