+++ src/Strategy/StrategyAwareTrait.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 namespace League\Route\Strategy;
 
+use League\Route\FreezeGuard;
+
 trait StrategyAwareTrait
 {
     protected ?StrategyInterface $strategy = null;
 
     public function setStrategy(StrategyInterface $strategy): StrategyAwareInterface
     {
+        FreezeGuard::assertNotFrozen($this);
         $this->strategy = $strategy;
         return $this;
     }